- [*] move Version 2 Function definitions from template.go to easygen.go
