+++ easygen.go
@@ -20,6 +20,7 @@ package easygen
 import (
 	"bytes"
 	"fmt"
+	"io"
 	"io/ioutil"
 	"os"
 	"path/filepath"
@@ -155,6 +156,49 @@ func ParseFiles(HTML bool, filenames ...string) (Template, error) {
 
 ////////////////////////////////////////////////////////////////////////////
 // Version 2 Function definitions
+
+// Process will process the standard easygen input: the `fileName` is for both template and data file names, and produce output from the template according to the corresponding driving data.
+func Process(t Template, wr io.Writer, fileName string) error {
+	return Process2(t, wr, fileName, fileName)
+}
+
+// Process2 will process the case that both template and data file names are given, and produce output according to the given template and driving data files,
+// specified via fileNameTempl and fileNames (for data) respectively.
+func Process2(t Template, wr io.Writer, fileNameTempl string, fileNames ...string) error {
+	fileName := fileNames[0]
+
+	source, err := ioutil.ReadFile(fileName + Opts.ExtYaml)
+	checkError(err)
+
+	m := make(map[interface{}]interface{})
+
+	err = yaml.Unmarshal(source, &m)
+	checkError(err)
+
+	env := make(map[string]string)
+	for _, e := range os.Environ() {
+		sep := strings.Index(e, "=")
+		env[e[0:sep]] = e[sep+1:]
+	}
+	m["ENV"] = env
+	//fmt.Printf("] %+v\n", m)
+
+	// template file name
+	fileNameT := fileNameTempl
+	if len(Opts.TemplateFile) > 0 {
+		fileNameT = Opts.TemplateFile
+	}
+	fileNameT = fileNameT + Opts.ExtTmpl
+	tn, err := t.ParseFiles(fileNameT)
+	checkError(err)
+
+	return tn.ExecuteTemplate(wr, filepath.Base(fileNameT), m)
+}
+
+// Process0 will produce output according to the driving data *without* a template file, using the string from strTempl as the template
+func Process0(t Template, wr io.Writer, strTempl string, fileNames ...string) error {
+	return nil
+}
 
 ////////////////////////////////////////////////////////////////////////////
 // Support Function definitions
